Bugfix: also support the back button on /locations and /reviews
