+++ src/components/connect/ConnectMap.js
@@ -1,9 +1,9 @@
 import { useEffect } from 'react'
 import { useDispatch, useSelector } from 'react-redux'
-import { useLocation } from 'react-router-dom'
 
 import { setInitialView } from '../../redux/mapSlice'
 import { parseCurrentUrl } from '../../utils/appUrl'
+import ConnectPath from './ConnectPath'
 
 const DEFAULT_LAT = 40.1125785
 const DEFAULT_LNG = -88.2287926
@@ -11,8 +11,7 @@ const DEFAULT_ZOOM = 4
 
 const ConnectMap = () => {
   const dispatch = useDispatch()
-  const { initialView, googleMap } = useSelector((state) => state.map)
-  const { pathname } = useLocation()
+  const { initialView } = useSelector((state) => state.map)
   const hasInitialView = !!initialView
   const parsedUrl = parseCurrentUrl()
   const view = parsedUrl.view || {
@@ -26,25 +25,7 @@ const ConnectMap = () => {
     }
   }, [dispatch, hasInitialView]) //eslint-disable-line
 
-  useEffect(() => {
-    if (hasInitialView && googleMap) {
-      const currentCenter = googleMap.getCenter()
-      const currentZoom = googleMap.getZoom()
-
-      if (
-        Math.abs(currentCenter.lat(), view.center.lat) > 1e-7 ||
-        Math.abs(currentCenter.lng(), view.center.lng) > 1e-7
-      ) {
-        googleMap.setCenter(view.center)
-      }
-
-      if (currentZoom !== view.zoom) {
-        googleMap.setZoom(view.zoom)
-      }
-    }
-  }, [pathname]) //eslint-disable-line
-
-  return null
+  return <ConnectPath />
 }
 
 export default ConnectMap
