+++ src/components/connect/ConnectPath.js
@@ -1,0 +1,34 @@
+import { useEffect } from 'react'
+import { useSelector } from 'react-redux'
+import { useLocation } from 'react-router-dom'
+
+import { parseCurrentUrl } from '../../utils/appUrl'
+
+const ConnectPath = () => {
+  const { initialView, googleMap } = useSelector((state) => state.map)
+  const { pathname } = useLocation()
+  const hasInitialView = !!initialView
+  const { view } = parseCurrentUrl()
+
+  useEffect(() => {
+    if (hasInitialView && googleMap && view) {
+      const currentCenter = googleMap.getCenter()
+      const currentZoom = googleMap.getZoom()
+
+      if (
+        Math.abs(currentCenter.lat(), view.center.lat) > 1e-7 ||
+        Math.abs(currentCenter.lng(), view.center.lng) > 1e-7
+      ) {
+        googleMap.setCenter(view.center)
+      }
+
+      if (currentZoom !== view.zoom) {
+        googleMap.setZoom(view.zoom)
+      }
+    }
+  }, [pathname, hasInitialView, googleMap]) //eslint-disable-line
+
+  return null
+}
+
+export default ConnectPath
